Everything, except messageLogger is working now, ready to merge
